refactor(ui): update imports to use new component paths and remove unused code
